cleanup code smells
